move poll code to app
