--- poll/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from . models import Poll, Question
-
-# Register your models here.
-
-admin.site.register(Poll)
-admin.site.register(Question)
